computor-testing/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "computor-testing"
version = "0.1.0"
description = "Computor Testing Framework - Supports Octave, R, Python, C/C++, Fortran, Julia, Document/Text testing"
authors = [
  { name = "Maximilian Mandlez", email = "maximilian.mandlez@tugraz.at" },
]
dependencies = [
  "computor-types>=0.1.0",
  "click>=8.1.7",
  "colorama>=0.4.6",
  "matplotlib>=3.8.2",
  "numpy>=1.26.2",
  "pydantic>=2.5.2",
  "pytest>=7.4.3",
  "pytest-metadata>=3.1.1",
  "PyYAML>=6.0.1",
  "stopit>=1.1.2",
  "scipy>=1.11.0",
  "sympy>=1.12",
  "uncertainties>=3.1.7",
]
readme = 'README.md'
requires-python = ">=3.10, <4"
license = { file = 'LICENSE' }
classifiers = [
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
]

[project.scripts]
# Unified CLI
computor-test = "testers.cli:cli"

# Language-specific CLIs
octester = "testers.cli:octester_cli"
rtester = "testers.cli:rtester_cli"
pytester = "testers.cli:pytester_cli"
ctester = "testers.cli:ctester_cli"
ftester = "testers.cli:ftester_cli"
jltester = "testers.cli:jltester_cli"
doctester = "testers.cli:doctester_cli"

# Utilities
deps-installer = "dependencies.installer:cli"
blocks = "blocks.cli:cli"
sandbox = "sandbox.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["ctcore*", "ctexec*", "testers*", "dependencies*", "blocks*", "sandbox*"]
